refactor: redis 저장까지 구현
