+++ src/main/java/com/team1/spreet/controller/EmailController.java
@@ -1,14 +1,15 @@
 package com.team1.spreet.controller;
 
 import com.team1.spreet.dto.CustomResponseBody;
-import com.team1.spreet.dto.EmailDto;
 import com.team1.spreet.exception.SuccessStatusCode;
 import com.team1.spreet.service.EmailService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @Api(tags = "email")
 @RestController
@@ -24,10 +25,10 @@ public class EmailController {
         return new CustomResponseBody<>(SuccessStatusCode.EMAIL_SEND_SUCCESS);
     }
 
-    @ApiOperation(value = "이메일 인증 API")
-    @PostMapping("/api/user/confirm-email")
-    public CustomResponseBody emailConfirm(@RequestBody @ApiParam(value = "이메일 인증을 위한 정보") EmailDto emailDto) throws Exception{
-        emailService.emailConfirm(emailDto);
-        return new CustomResponseBody<>(SuccessStatusCode.EMAIL_CONFIRM_SUCCESS);
-    }
+//    @ApiOperation(value = "이메일 인증 API")
+//    @PostMapping("/api/user/confirm-email")
+//    public CustomResponseBody emailConfirm(@RequestBody @ApiParam(value = "이메일 인증을 위한 정보") EmailDto emailDto) throws Exception{
+//        emailService.emailConfirm(emailDto);
+//        return new CustomResponseBody<>(SuccessStatusCode.EMAIL_CONFIRM_SUCCESS);
+//    }
 }
